# week4_library_demo/openAPI.yaml
openapi: 3.0.0
info:
  title: Library API Gateway
  version: "1.0.0"
  description: |
    API Gateway trung gian giữa Frontend và các microservice:
    - **Auth Service**: Xác thực, phân quyền, OAuth2
    - **Book Service**: Quản lý sách
    - **Transaction Service**: Giao dịch mượn/trả sách
servers:
  - url: http://localhost:8080/api
    description: Local API Gateway Server

tags:
  - name: Auth
    description: Xác thực người dùng (JWT, OAuth2, Google)
  - name: Books
    description: Quản lý tài nguyên sách
  - name: Transactions
    description: Quản lý giao dịch mượn / trả sách

paths:
  /users:
    post:
      tags: [Auth]
      summary: Đăng ký người dùng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Tạo người dùng thành công
        "400":
          description: Thiếu thông tin bắt buộc

  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập (tạo access token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: Trả về access token và thông tin user
        "401":
          description: Sai thông tin đăng nhập

  /auth/refresh-token:
    put:
      tags: [Auth]
      summary: Làm mới access token bằng refresh token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token mới được cấp
        "401":
          description: Token không hợp lệ hoặc hết hạn

  /auth/logout:
    delete:
      tags: [Auth]
      summary: Đăng xuất người dùng hiện tại
      description: |
        Hủy hiệu lực của token hiện tại.  
        Refresh token được gửi kèm trong **HTTP Cookie**.
        - Access token được xác thực qua header: `Authorization: Bearer <token>`
        - Refresh token nằm trong cookie (ví dụ: `refresh_token=<jwt>`)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Đăng xuất thành công (cookie bị xóa)
        "401":
          description: Token không hợp lệ hoặc đã hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/google/login:
    get:
      tags: [Auth]
      summary: Đăng nhập bằng Google OAuth2
      responses:
        "302":
          description: Redirect tới trang đăng nhập Google

  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Callback từ Google OAuth2
      responses:
        "200":
          description: Đăng nhập thành công và trả về token

  /users/nplus1:
    get:
      tags: [Auth]
      summary: (Admin) Debug N+1 Query Problem
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dữ liệu minh họa cho vấn đề N+1

  /users/eager:
    get:
      tags: [Auth]
      summary: (Admin) Debug Eager Loading
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dữ liệu minh họa cho giải pháp Eager Loading

  /users/batch:
    get:
      tags: [Auth]
      summary: (Admin) Debug Batch Loading
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dữ liệu minh họa cho Batch Loading

  /books:
    get:
      tags: [Books]
      summary: Lấy danh sách sách (tự động offset hoặc cursor pagination)
      security:
        - bearerAuth: []
      description: |
        Hỗ trợ 2 loại phân trang:
        - **Offset**: dùng `page` và `limit`
        - **Cursor**: dùng `after_cursor` hoặc `before_cursor`
        - `author_filter`: lọc theo tên tác giả (partial match)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
          description: Số bản ghi tối đa mỗi trang
        - name: page
          in: query
          schema: { type: integer, default: 1 }
          description: Số trang (chỉ dùng khi phân trang kiểu offset)
        - name: after_cursor
          in: query
          schema: { type: string }
          description: Cursor để lấy trang kế tiếp
        - name: before_cursor
          in: query
          schema: { type: string }
          description: Cursor để lấy trang trước đó
        - name: author_filter
          in: query
          schema: { type: string }
          description: Lọc sách theo tên tác giả
      responses:
        "200":
          description: Danh sách sách trả về
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BooksOffsetResponse'
                  - $ref: '#/components/schemas/BooksCursorResponse'
        "304":
          description: Not Modified (ETag matched)
        "400":
          description: Tham số không hợp lệ
    post:
      tags: [Books]
      summary: (Admin) Tạo sách mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        "201":
          description: Tạo sách thành công
        "403":
          description: Cần quyền admin

  /books/{book_id}:
    parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Books]
      summary: Lấy chi tiết 1 sách
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thông tin sách
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        "304":
          description: Not Modified (ETag matched)
        "404":
          description: Không tìm thấy sách
    put:
      tags: [Books]
      summary: (Admin) Cập nhật thông tin sách
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdate'
      responses:
        "200":
          description: Cập nhật thành công
        "404":
          description: Không tìm thấy sách
    delete:
      tags: [Books]
      summary: (Admin) Xóa sách
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Xóa thành công
        "404":
          description: Không tìm thấy sách

  /transactions:
    post:
      tags: [Transactions]
      summary: Tạo giao dịch (mượn/trả sách)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id, quantity, type]
              properties:
                book_id: { type: integer }
                quantity: { type: integer }
                type: { type: string, example: "borrow" }
      responses:
        "201":
          description: Giao dịch được tạo thành công
        "400":
          description: Dữ liệu không hợp lệ

  /me/borrowed-books:
    get:
      tags: [Transactions]
      summary: Lấy danh sách sách đang mượn của user hiện tại
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách sách đang mượn
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BorrowedBook'

  /me/transactions:
    get:
      tags: [Transactions]
      summary: Lấy lịch sử giao dịch của user hiện tại
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách giao dịch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Book:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        author: { type: string }
        available_copies: { type: integer }
    BookCreate:
      type: object
      required: [title, author, available_copies]
      properties:
        title: { type: string }
        author: { type: string }
        available_copies: { type: integer }
    BookUpdate:
      type: object
      properties:
        title: { type: string }
        author: { type: string }
        available_copies: { type: integer }
    BooksOffsetResponse:
      type: object
      properties:
        pagination_type: { type: string, example: "offset" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/Book' }
    BooksCursorResponse:
      type: object
      properties:
        pagination_type: { type: string, example: "cursor" }
        items:
          type: array
          items: { $ref: '#/components/schemas/Book' }
        next_cursor: { type: string, nullable: true }
        prev_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
    Transaction:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        book_id: { type: integer }
        type: { type: string, enum: [borrow, return] }
        quantity: { type: integer }
        status: { type: string, example: "completed" }
    BorrowedBook:
      type: object
      properties:
        book:
          $ref: '#/components/schemas/Book'
        borrowed_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string